2013/10/01

+++ Eland/FXLabel.m
@@ -195,7 +195,7 @@
     UIFont *font = [self.font fontWithSize:fontSize];
 
     //set position
-    switch (self.textAlignment)
+    switch ((int)self.textAlignment)
     {
         case NSTextAlignmentCenter:
         {
